requirements.txt
google-genai
streamlit
apscheduler
